refactor: add routes constants
- Add storage to index.ts
- Update storage.ts
- Update sidebar.tsx use routes constants
- Update baseQueryWithReauth.ts
- Update useSignIn.ts
- Update useLogout.tsx
