+++ src/shared/config/routes.ts
@@ -1,11 +1,16 @@
 export const ROUTES = {
+  CREATE: '/create',
   CREATE_NEW_PASSWORD: '/auth/recovery',
+  FAVORITES: '/favorites',
   FORGOT_PASSWORD: '/auth/forgotPasswordPage',
   HOME: '/home',
+  MESSENGER: '/messenger',
   PRIVACY_POLICY: '/auth/privacyPolicyPage',
   PROFILE: '/profile',
+  SEARCH: '/search',
   SIGN_IN: '/auth/signInPage',
   SIGN_UP: '/auth/signUpPage',
+  STATICS: '/statics',
   TERMS_OF_SERVICE: '/auth/termsOfServicePage',
 }
