+++ src/shared/utils/index.ts
@@ -3,3 +3,4 @@ export * from './commonFormUtils'
 export * from './handleErrorResponse'
 export * from './hooks'
 export * from './isValidErrorResponse'
+export * from './storage'
